libc: minimal: errno.h, strerror to common, strerror_r to posix
The strerror_r() function is POSIX. It's not part of ISO C.
Moreover, the versions available in both Newlib and Picolibc do
not seem to comply with the POSIX specification OOTB.

Moving this function to lib/posix ensures that Zephyr
applications have a consistent implementation of the POSIX API.
