+++ doc/develop/languages/c/common_libc.rst
@@ -45,3 +45,27 @@ The standard dynamic memory management interface functions implemented by
 the common C library are thread safe and may be simultaneously called by
 multiple threads. These functions are implemented in
 :file:`lib/libc/common/source/stdlib/malloc.c`.
+
+Error numbers
+*************
+
+Error numbers are used throughout Zephyr APIs to signal error conditions as
+return values from functions. They are typically returned as the negative value
+of the integer literals defined in this section, and are defined in the
+:file:`errno.h` header file.
+
+A subset of the error numbers defined in the `POSIX errno.h specification`_ and
+other de-facto standard sources have been added to the common libc resources.
+
+A conscious effort is made in Zephyr to keep the values of the common libc
+error numbers consistent with the different implementations of the C standard
+libraries supported by Zephyr. The minimal libc :file:`errno.h` is checked
+against that of the :ref:`Newlib <c_library_newlib>` to ensure that the error
+numbers are kept aligned.
+
+Below is a list of the error number definitions. For the actual numeric values
+please refer to :file:`lib/libc/common/include/errno.h`.
+
+.. doxygengroup:: system_errno
+
+.. _`POSIX errno.h specification`: https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/errno.h.html
