+++ lib/CMakeLists.txt
@@ -4,9 +4,7 @@
 add_compile_options($<TARGET_PROPERTY:compiler,warning_shadow_variables>)
 
 add_subdirectory(crc)
-if(NOT CONFIG_EXTERNAL_LIBC)
 add_subdirectory(libc)
-endif()
 if(NOT CONFIG_NATIVE_LIBC)
 add_subdirectory(posix)
 endif()
