+++ lib/libc/Kconfig
@@ -74,6 +74,7 @@ config MINIMAL_LIBC
 	imply COMPILER_FREESTANDING
 	select COMMON_LIBC_ABORT
 	select COMMON_LIBC_STRNLEN
+	select COMMON_LIBC_STRERROR
 	help
 	  Build with minimal C library.
 
@@ -86,6 +87,7 @@ config PICOLIBC
 	imply COMMON_LIBC_MALLOC
 	depends on !NATIVE_APPLICATION
 	depends on PICOLIBC_SUPPORTED
+	select COMMON_LIBC_STRERROR
 	help
 	  Build with picolibc library. The picolibc library is built as
 	  a module if PICOLIBC_MODULE is set, otherwise picolibc is
@@ -97,6 +99,7 @@ config NEWLIB_LIBC
 	depends on !NATIVE_APPLICATION
 	depends on NEWLIB_LIBC_SUPPORTED
 	select NEED_LIBC_MEM_PARTITION
+	select COMMON_LIBC_STRERROR
 	help
 	  Build with newlib library. The newlib library is expected to be
 	  part of the SDK in this case.
