+++ lib/libc/minimal/Kconfig
@@ -90,14 +90,11 @@ config MINIMAL_LIBC_TIME
 	  to set CONFIG_MINIMAL_LIBC_NON_REENTRANT_FUNCTIONS=y.
 
 config MINIMAL_LIBC_STRING_ERROR_TABLE
-	bool "String error table for strerror() and strerror_r()"
+	bool "String error table for strerror() [DEPRECATED]"
+	select DEPRECATED
+	select COMMON_LIBC_STRING_ERROR_TABLE
 	help
-	  Select this option to ensure that streror(), strerror_r()
-	  produce strings corresponding to the descriptions in errno.h.
-
-	  The string error table can add ~2kiB to ROM. As such, it is
-	  disabled by default. In this case, strerror() and strerror_r()
-	  symbols are still present, but the functions produce an empty
-	  string.
+	  This option is deprecated. Use CONFIG_COMMON_LIBC_STRING_ERROR_TABLE
+	  instead.
 
 endif # MINIMAL_LIBC
