+++ lib/posix/Kconfig
@@ -54,3 +54,6 @@ source "lib/posix/Kconfig.signal"
 source "lib/posix/Kconfig.spinlock"
 source "lib/posix/Kconfig.timer"
 source "lib/posix/Kconfig.uname"
+
+# POSIX extensions to ISO C
+source "lib/posix/Kconfig.cx.string"
