+++ scripts/ci/errno.py
@@ -4,9 +4,9 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Check minimal libc error numbers against newlib.
+"""Check common libc error numbers against newlib.
 
-This script loads the errno.h included in Zephyr's minimal libc and checks its
+This script loads the errno.h included in Zephyr's common libc and checks its
 contents against the SDK's newlib errno.h. This is done to ensure that both C
 libraries are aligned at all times.
 """
@@ -30,20 +30,20 @@ def parse_errno(path):
 
 def main():
 
-    minimal = Path("lib/libc/minimal/include/errno.h")
+    common = Path("lib/libc/common/include/errno.h")
     newlib = Path("arm-zephyr-eabi/arm-zephyr-eabi/include/sys/errno.h")
 
     try:
-        minimal = os.environ['ZEPHYR_BASE'] / minimal
+        common = os.environ['ZEPHYR_BASE'] / common
         newlib = os.environ['ZEPHYR_SDK_INSTALL_DIR'] / newlib
     except KeyError as e:
         print(f'Environment variable missing: {e}', file=sys.stderr)
         sys.exit(1)
 
-    minimal = parse_errno(minimal)
+    common = parse_errno(common)
     newlib = parse_errno(newlib)
 
-    for e in minimal:
+    for e in common:
         if e[0] not in [x[0] for x in newlib] or e[1] != next(
             filter(lambda _e: _e[0] == e[0], newlib))[1]:
             print('Invalid entry in errno.h:', file=sys.stderr)
