+++ tests/lib/c_lib/strerror/src/main.c
@@ -4,10 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifdef CONFIG_NEWLIB_LIBC
-#define _POSIX_C_SOURCE 200809
-#endif
-
 #include <errno.h>
 #include <string.h>
 
@@ -19,7 +15,7 @@ ZTEST(libc_strerror, test_strerror)
 	const char *actual;
 
 	errno = 4242;
-	if (IS_ENABLED(CONFIG_MINIMAL_LIBC_DISABLE_STRING_ERROR_TABLE)) {
+	if (!IS_ENABLED(CONFIG_COMMON_LIBC_STRING_ERROR_TABLE)) {
 		expected = "";
 		actual = strerror(EINVAL);
 	} else {
@@ -47,54 +43,12 @@ ZTEST(libc_strerror, test_strerror)
 #endif
 
 	/* consistent behaviour for "Success" */
-	if (!IS_ENABLED(CONFIG_MINIMAL_LIBC_DISABLE_STRING_ERROR_TABLE)) {
+	if (IS_ENABLED(CONFIG_COMMON_LIBC_STRING_ERROR_TABLE)) {
 		expected = "Success";
 		actual = strerror(0);
 		zassert_equal(0, strcmp(expected, actual),
 			      "mismatch: exp: %s act: %s", expected, actual);
 	}
-}
-
-ZTEST(libc_strerror, test_strerror_r)
-{
-	const char *expected;
-	char actual[] = {'1', 'n', 'v', 'a', '1', '1', 'd', ' ',  'a',
-			 '2', 'g', 'u', 'm', '3', 'n', '7', 0x00, 0x42};
-	static const size_t n = sizeof(actual);
-
-	if (IS_ENABLED(CONFIG_NEWLIB_LIBC) || IS_ENABLED(CONFIG_ARCMWDT_LIBC)) {
-		/* FIXME: Please see Issue #46846 */
-		ztest_test_skip();
-	}
-
-	errno = 4242;
-	if (IS_ENABLED(CONFIG_MINIMAL_LIBC_DISABLE_STRING_ERROR_TABLE)) {
-		expected = "";
-		zassert_equal(0, strerror_r(EINVAL, actual, n), "");
-		zassert_equal(0, strncmp(expected, actual, n),
-			      "mismatch: exp: %s act: %s", expected, actual);
-	} else {
-		expected = "Invalid argument";
-		zassert_equal(0, strerror_r(EINVAL, actual, n), "%d",
-			      strerror_r(EINVAL, actual, n));
-		zassert_equal(0, strncmp(expected, actual, n),
-			      "mismatch: exp: %s act: %s", expected, actual);
-		/* only the necessary buffer area is written */
-		zassert_equal(0x42, (uint8_t)actual[n - 1],
-			      "exp: %02x act: %02x", 0x42,
-			      (uint8_t)actual[n - 1]);
-
-		zassert_equal(ERANGE, strerror_r(EINVAL, actual, 0), "");
-	}
-
-	/* do not change errno on success */
-	zassert_equal(4242, errno, "");
-
-	errno = 0;
-	zassert_equal(EINVAL, strerror_r(-42, actual, n), "");
-	zassert_equal(EINVAL, strerror_r(4242, actual, n), "");
-	/* do not change errno on failure */
-	zassert_equal(0, errno, "");
 }
 
 ZTEST_SUITE(libc_strerror, NULL, NULL, NULL, NULL, NULL);
