+++ tests/posix/cx/src/string/_main.c
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2023, Meta
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <zephyr/ztest.h>
+
+ZTEST_SUITE(posix_cx_string, NULL, NULL, NULL, NULL, NULL);
